% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_ollama.R
\name{ask_ollama_vision}
\alias{ask_ollama_vision}
\title{Interroger le Modèle Llama3.2-Vision via Ollama}
\usage{
ask_ollama_vision(
  question,
  image_path = NULL,
  model_name = "llama3.2-vision",
  base_url = "https://ollama-clem.lab.sspcloud.fr"
)
}
\arguments{
\item{question}{character La question à poser sur l'image}

\item{image_path}{character Chemin vers l'image à analyser}

\item{model_name}{character Nom du modèle (défaut: "llama3.2-vision")}
}
\value{
character La réponse du modèle
}
\description{
Interroger le Modèle Llama3.2-Vision via Ollama
}
\examples{
\dontrun{
reponse <- ask_ollama_vision(
  question = "What is in this image?",
  image_path = "image.jpg"
)
}
}
