% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lodel_avaibility.R
\name{check_model_availability}
\alias{check_model_availability}
\title{Vérifier la Disponibilité d'un Modèle via une API}
\usage{
check_model_availability(
  model_name,
  base_url = "https://ollama-clem.lab.sspcloud.fr"
)
}
\arguments{
\item{model_name}{\code{character} : Le nom du modèle à vérifier.}

\item{base_url}{\code{character} : L'URL de base de l'API à interroger. Par défaut, \code{"https://ollama-clem.lab.sspcloud.fr"}.}
}
\value{
\code{logical} : \code{TRUE} si le modèle est disponible, \code{FALSE} sinon.
}
\description{
Cette fonction vérifie si un modèle spécifique est disponible en interrogeant une API donnée. Elle envoie une requête GET à l'URL de base spécifiée, récupère la liste des modèles disponibles, et détermine si le modèle souhaité est présent dans cette liste.
}
\details{
La fonction effectue les opérations suivantes :
\itemize{
\item Envoie une requête GET à \code{base_url}/api/tags pour récupérer les informations des modèles disponibles.
\item Vérifie si la requête a réussi (code de statut HTTP 200).
\item Parse le contenu de la réponse pour extraire les noms des modèles disponibles, en retirant le suffixe \code{:latest} si présent.
\item Vérifie si \code{model_name} est présent dans la liste des noms de modèles extraits.
}

Si la requête échoue ou si le modèle n'est pas trouvé, la fonction retourne \code{FALSE}.
}
\examples{
\dontrun{
# Vérifier la disponibilité d'un modèle nommé "gpt-4"
disponible <- check_model_availability("gpt-4")
if (disponible) {
  message("Le modèle 'gpt-4' est disponible.")
} else {
  message("Le modèle 'gpt-4' n'est pas disponible.")
}

# Vérifier la disponibilité d'un modèle avec une URL de base personnalisée
disponible <- check_model_availability("gpt-4", base_url = "https://custom-api.example.com")
}

}
