% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_ollama.R
\name{process_llava_response}
\alias{process_llava_response}
\title{Traiter la réponse de l'API LLaVA}
\usage{
process_llava_response(response)
}
\arguments{
\item{response}{character Vector contenant les lignes de réponse JSON de LLaVA}
}
\value{
character Le texte assemblé de la réponse
}
\description{
Traiter la réponse de l'API LLaVA
}
\keyword{internal}
